requirements.txt
reportlab==4.0.9
streamlit
pandas
numpy
pillow
packaging
reportlab
streamlit-drawable-canvas-fix